:sparkles: GET /items?kind=
