--- lambdas/getItem/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "@subterra/lambdas-get-item",
-  "version": "1.0.0",
-  "main": "index.js",
-  "private": true,
-  "arn": "arn:aws:lambda:eu-west-3:427962677004:function:subterra_getItem",
-  "dependencies": {},
-  "workspaceDependencies": {}
-}
