--- lambdas/getItem/src/index.js
@@ -1,8 +1,0 @@
-export const handler = async (event) => {
-  // TODO: check autorization? With a lambda auth from api gateway maybe??
-
-  return {
-    statusCode: 200,
-    body: 'todo',
-  }
-}
