+++ lambdas/getItems/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@subterra/lambdas-get-items",
+  "version": "1.0.0",
+  "main": "index.js",
+  "private": true,
+  "arn": "arn:aws:lambda:eu-west-3:427962677004:function:subterra_getItems",
+  "dependencies": {
+    "@fabienjuif/dynamo-client": "^2.0.0"
+  },
+  "workspaceDependencies": {
+    "@subterra/rest-utils": "^1.0.0"
+  }
+}
