+++ packages/rest-utils/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@subterra/rest-utils",
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT",
+  "private": true
+}
